fix(dialogue manager): fix typo
