+++ Assets/Scripts/Dialog/DialogueManagerV2.cs.meta
